Creating 2 spaces, and then opening the first object's tweak bar in both spaces no longer crashes. woo

+++ Dev/StaticSheepNeoBuildSystem/game/SheepEngine/engine/core/EngineTweak.cpp
@@ -139,7 +139,7 @@ namespace Framework
   static void EditorPlayLevel(void* clientData)
   {
     std::vector<GameSpace*>& gameSpaces = ENGINE->Spaces();
-    for (int i = 0; i < gameSpaces.size(); ++i)
+    for (size_t i = 0; i < gameSpaces.size(); ++i)
     {
       //FACTORY->SaveSpaceToFile(gameSpaces[i], "test", true);
     }
